+++ changelog.md
@@ -20,3 +20,4 @@ See https://modrinth.com/mod/ad-astra/version/1.15.7 for the full changelog.
 - Fixed random accessed from netty thread crash when orbit sounds play. (#473).
 - Fixed The Ti-69s red and green lights being swapped from the item texture (#479).
 - Added painting translations (#485).
+- The Nasa workbench now works with recipes with less than 14 ingredients (#483).